+++ ai.symfony.com/assets/styles/app.css
@@ -215,10 +215,10 @@ button#themeToggle {
     color: white;
 }
 
-.demo-audio .demo-icon { background: linear-gradient(180deg, #42DEEE, #7069B0); }
 .demo-blog .demo-icon { background: linear-gradient(180deg, #433F77, #C43BC2); }
 .demo-crop .demo-icon { background: linear-gradient(180deg, #85A72B, #97BC43); }
 .demo-recipe .demo-icon { background: linear-gradient(180deg, #83A659, #71BCB8); }
+.demo-speech .demo-icon { background: linear-gradient(180deg, #42DEEE, #7069B0); }
 .demo-turbo .demo-icon { background: linear-gradient(180deg, #E94E77, #D68189); }
 .demo-video .demo-icon { background: linear-gradient(180deg, #3B9D87, #35A781); }
 .demo-wikipedia .demo-icon { background: linear-gradient(180deg, #1CA574, #56AB48); }
