+++ ai.symfony.com/templates/homepage.html.twig
@@ -256,7 +256,7 @@
                     </div>
 
                     <div class="col-md-6">
-                        <article class="demo-audio sf-ai-card sf-ai-card-hover h-100">
+                        <article class="demo-speech sf-ai-card sf-ai-card-hover h-100">
                             <div class="position-relative card-body p-3 p-lg-4 d-flex gap-3 align-items-start">
                                 <div class="demo-icon bg-success bg-opacity-10 text-success flex-shrink-0">
                                     {{ ux_icon('tabler:microphone', {width: 42, height: 42}) }}
@@ -264,7 +264,7 @@
                                 <div class="flex-grow-1">
                                     <div class="d-flex align-items-start gap-2 mb-1">
                                         <h3 class="h5 ff-title fw-bold mb-0">
-                                            <a href="https://github.com/symfony/ai-demo" class="stretched-link">Audio Bot + Subagent</a>
+                                            <a href="https://github.com/symfony/ai-demo" class="stretched-link">Speech Bot + Subagent</a>
                                         </h3>
                                         {{ ux_icon('tabler:arrow-right', {width: 24, height: 24, class: 'ms-1'}) }}
                                     </div>
