+++ demo/assets/app.js
@@ -1,10 +1,10 @@
 import './bootstrap.js';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './styles/app.css';
-import './styles/audio.css';
 import './styles/blog.css';
 import './styles/crop.css';
 import './styles/recipe.css';
+import './styles/speech.css';
 import './styles/stream.css';
 import './styles/youtube.css';
 import './styles/video.css';
