+++ demo/assets/styles/speech.css
@@ -1,4 +1,4 @@
-.audio {
+.speech {
     body&, .card-img-top {
         background: #df662f;
         background: linear-gradient(0deg, #df662f 0%, #a80a1d 100%);
