+++ demo/config/routes.yaml
@@ -4,13 +4,6 @@ index:
     defaults:
         template:  'index.html.twig'
 
-audio:
-    path: '/audio'
-    controller: 'Symfony\Bundle\FrameworkBundle\Controller\TemplateController'
-    defaults:
-        template:  'chat.html.twig'
-        context: { chat: 'audio' }
-
 blog:
     path: '/blog'
     controller: 'Symfony\Bundle\FrameworkBundle\Controller\TemplateController'
@@ -30,6 +23,13 @@ recipe:
     defaults:
         template:  'chat.html.twig'
         context: { chat: 'recipe' }
+
+speech:
+    path: '/speech'
+    controller: 'Symfony\Bundle\FrameworkBundle\Controller\TemplateController'
+    defaults:
+        template:  'chat.html.twig'
+        context: { chat: 'speech' }
 
 stream:
     path: '/stream'
