+++ demo/templates/components/speech.html.twig
@@ -1,9 +1,9 @@
 {% import "_message.html.twig" as msg %}
 
-<div class="card mx-auto shadow-lg" {{ attributes.defaults(stimulus_controller('audio')) }}>
+<div class="card mx-auto shadow-lg" {{ attributes.defaults(stimulus_controller('speech')) }}>
     <div class="card-header p-2">
         {{ ux_icon('iconoir:microphone-solid', { height: '32px', width: '32px' }) }}
-        <strong class="ms-1 pt-1 d-inline-block">Conversational Bot</strong>
+        <strong class="ms-1 pt-1 d-inline-block">Speech Bot</strong>
         <button id="chat-reset" class="btn btn-sm btn-outline-secondary float-end">{{ ux_icon('material-symbols:cancel') }} Reset Chat</button>
     </div>
     <div id="chat-body" class="card-body p-4 overflow-auto">
@@ -16,14 +16,14 @@
                         {{ ux_icon('fluent:bot-24-filled', { height: '45px', width: '45px' }) }}
                     </div>
                     <div class="ps-2">
-                        <audio class="pt-3" controls {{ loop.last ? 'autoplay' }} src="{{ message.metadata.get('audio') }}"></audio>
+                        <audio class="pt-3" controls {{ loop.last ? 'autoplay' }} src="{{ message.metadata.get('speech') }}"></audio>
                     </div>
                 </div>
             {% endif %}
         {% else %}
             <div id="welcome" class="text-center mt-5 py-5 bg-white rounded-5 shadow-sm w-75 mx-auto">
                 {{ ux_icon('iconoir:microphone-solid', { height: '200px', width: '200px' }) }}
-                <h4 class="mt-5">Audio Bot</h4>
+                <h4 class="mt-5">Speech Bot</h4>
                 <span class="text-muted">Please hit the button below to start talking and again to stop</span>
             </div>
         {% endfor %}
